compat/historical.html

<!doctype html>
<title>Historical non-standard features</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_false("getMatchedCSSRules" in window);
}, "getMatchedCSSRules() should not exist");
</script>
